Fix crowdfunding summary endpoint and document data flow
